+++ src/main/java/study/racingcar/domain/Awards.java
@@ -1,4 +1,4 @@
-package study.racingcar.model;
+package study.racingcar.domain;
 
 import study.racingcar.strategy.AwardsStrategy;
