+++ src/main/java/study/racingcar/domain/Cars.java
@@ -1,4 +1,4 @@
-package study.racingcar.model;
+package study.racingcar.domain;
 
 import study.racingcar.strategy.MovableStrategy;
 import study.racingcar.strategy.RandomMovableStrategy;
