+++ src/main/java/study/racingcar/domain/RacingGameInfo.java
@@ -1,4 +1,4 @@
-package study.racingcar.model;
+package study.racingcar.domain;
 
 import java.util.Set;
