+++ src/main/java/study/racingcar/view/input/InputView.java
@@ -1,6 +1,6 @@
-package study.racingcar.input;
+package study.racingcar.view.input;
 
-import study.racingcar.model.RacingGameInfo;
+import study.racingcar.domain.RacingGameInfo;
 
 /**
  * Created by wyparks2@gmail.com on 2019-06-17
