+++ src/main/java/study/racingcar/view/output/ResultView.java
@@ -1,7 +1,7 @@
-package study.racingcar.output;
+package study.racingcar.view.output;
 
-import study.racingcar.model.Cars;
-import study.racingcar.model.CarsRacingLog;
+import study.racingcar.domain.Cars;
+import study.racingcar.domain.CarsRacingLog;
 
 import java.util.List;
