+++ src/test/java/study/racingcar/domain/CarsTest.java
@@ -1,4 +1,4 @@
-package study.racingcar.model;
+package study.racingcar.domain;
 
 import org.junit.jupiter.api.Test;
 import study.racingcar.strategy.ManualMovableStrategy;
